requirements.txt
Flask>=3.0,<4.0
Flask-Login>=0.6,<1.0
Flask-SQLAlchemy>=3.1,<4.0
Flask-WTF>=1.2,<2.0
python-dotenv>=1.0,<2.0
requests>=2.32,<3.0
psycopg[binary]>=3.2,<4.0
waitress>=2.2,<3.0
waitress>=2.2,<3.0
gunicorn>=21.2,<22.0
